Late binding of the frame in subscription/cofx/fx
Make sure subscriptions and events use the frame that is passed to
`subscribe`/`dispatch`, and not the frame that they are registered on, so that
it is possible to re-use registries across frames.

This also removes re-frame.db, since it's now vestigial, and removes
re-frame.registrar, since that has been reincarnated into re-frame.registry.