Provide a full top-level API replacement in re-frame.context
When using the React-context approach which allows having a separate frame (or
db, sub cache, etc) per React render tree, you should use functions which all
defer to the current context-set frame, rather than the default frame. To make
it harder to get this wrong re-frame.context as a namespace now is a full
API-level replacement for re-frame.core, so you can require that and mostly just
be fine.