Make extra keys possible in the frame, give frames an incremental id
If you want to keep extra state in the frame then it needs to be set at creation
time, so the reference that is given to the event-queue has the extra keys. If
you assoc later onto it the extra bits won't be available to event/fx handlers.

Add an incremental :frame-id for easy debugging.